Refactor App component: move implementation from app.rs to main.rs
